Allow alpha blending and subpixel rendering
The core of the library has been heavily revised to allow for this.
Changes:

- Latch bytes are no longer stored in the pixels array, saving a tiny
bit of memory.

- Use standard 8/8/8 RGB color instead of 7/7/7 GRB.  This allows the
blending to be done with 8-bit precision instead of 7, and makes it
possible to specify colors as standard RGB hex values.  This makes the
"Color" method mostly unnecessary.

- RGB is rearranged to GRB when a pixel is written to the pixels
array.

- Color bytes are reduced to 7 bits (and have the high bit set) when
they're sent to the LPD8806, not before. This is done using a gamma
lookup table stored in PROGMEM, borrowed from the LEDbeltKit_alt
example.

The following methods are added:

- blendPixel: Blends a pixel on top of an existing pixel.  This uses
the "screen" blend, which means that pixels are brightened only.

- renderSubPixel: Renders a blended 'subpixel' between two physical
pixels.  It does this using two blendPixel operations.

- clear: A simple method to clear the pixel memory (does not update the display).